split off data frame publishers
